Refactor CorrelationId
