+++ framework/src/Volo.Abp.AspNetCore/Volo/Abp/AspNetCore/Tracing/AspNetCoreCorrelationIdOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace Volo.Abp.AspNetCore.Tracing
+{
+    public class AspNetCoreCorrelationIdOptions
+    {
+        public string HeaderName { get; set; } = "X-Correlation-Id";
+    }
+}
