+++ framework/src/Volo.Abp.Core/Volo/Abp/Tracing/NullCorrelationIdProvider.cs
@@ -1,6 +1,6 @@
 ﻿using Volo.Abp.DependencyInjection;
 
-namespace Volo.Abp.Auditing
+namespace Volo.Abp.Tracing
 {
     public class NullCorrelationIdProvider : ICorrelationIdProvider, ISingletonDependency
     {
